HW10 Добавлена работа с Stream API в AnimalRepository и   дополнительные методы.
