+++ createAnimal-starter/src/main/java/ru/mts/hw10/Main.java
@@ -1,0 +1,11 @@
+package ru.mts.hw10;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        SpringApplication.run(Main.class, args);
+    }
+}
