+++ createAnimal-starter/src/main/java/ru/mts/hw10/factory/AnimalTypes.java
@@ -1,0 +1,9 @@
+package ru.mts.hw10.factory;
+
+public enum AnimalTypes {
+    WOLF,
+    SHARK,
+    DOG,
+    CAT,
+    UNDEFINED_ANIMAL
+}
