--- createAnimal-starter/src/main/java/ru/mts/hw9/Main.java
@@ -1,11 +1,0 @@
-package ru.mts.hw9;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-    public static void main(String[] args) {
-        SpringApplication.run(Main.class, args);
-    }
-}
