--- createAnimal-starter/src/main/java/ru/mts/hw9/factory/AnimalTypes.java
@@ -1,9 +1,0 @@
-package ru.mts.hw9.factory;
-
-public enum AnimalTypes {
-    WOLF,
-    SHARK,
-    DOG,
-    CAT,
-    UNDEFINED_ANIMAL
-}
